Why do this?
I live on a street where the speed limit is 30 MPH.  Many drivers drive much faster than this...but how can I prove this?  I would like to position my RaspberryPi with its camera on a light pole near the road to observe traffic and calculate vehicle speed.  A device such as this would be undetectable by radar detectors, so a truer measure of typical speeds would be produced.  Knowing the true speeds throughout the day would enable the police to determine the best time of day to position their vehicles to catch speeders or simply reduce speeds.